Remove repetitive code with a new component
