+++ frontend/src/pages/CreateFolder.js
@@ -1,23 +1,12 @@
-import React, { useState, Component } from "react";
-import Header from "../main-components/Header";
-import Sidebar from "../main-components/Sidebar";
+import React from "react";
 import CreateFolderContent from "../main-components/CreateFolderContent";
+import ContentWithSidebar from "./ContentWithSidebar";
 
 function CreateFolder() {
-  // Handles mobile menu button and updates sidebar view
-   const [sidebarVisibility, setSidebarVisibility] = useState(false);
-   const handleClick = (e) => {
-     setSidebarVisibility((sidebarVisibility) => !sidebarVisibility);
-   };
-
   return (
-    <div className="App">
-      <Header id="head" handleClick={handleClick}></Header>
-      <div id="main">
-        <Sidebar visible={sidebarVisibility}></Sidebar>
-        <CreateFolderContent></CreateFolderContent>  
-      </div>
-    </div>
+    <ContentWithSidebar>
+      <CreateFolderContent></CreateFolderContent>
+    </ContentWithSidebar>
   );
 }
