+++ frontend/src/pages/StudyMode.js
@@ -1,26 +1,15 @@
-import React, { useState, Component } from "react";
-import Header from "../main-components/Header";
-import Sidebar from "../main-components/Sidebar";
+import React from "react";
 import MyFoldersContent from "../main-components/MyFoldersContent";
+import ContentWithSidebar from "./ContentWithSidebar";
 
 function StudyMode() {
-  // Handles mobile menu button and updates sidebar view
-  const [sidebarVisibility, setSidebarVisibility] = useState(false);
-  const handleClick = (e) => {
-    setSidebarVisibility((sidebarVisibility) => !sidebarVisibility);
-  };
-
   return (
-    <div className='App'>
-      <Header id='head' handleClick={handleClick}></Header>
-      <div id='main'>
-        <Sidebar visible={sidebarVisibility}></Sidebar>
-        <MyFoldersContent
-          mainURL='/InsideStudyMode'
-          headingText='Select a folder to begin studying'>
-        </MyFoldersContent>
-      </div>
-    </div>
+    <ContentWithSidebar>
+      <MyFoldersContent
+        mainURL='/InsideStudyMode'
+        headingText='Select a folder to begin studying'>
+      </MyFoldersContent>
+    </ContentWithSidebar>
   );
 }
